divide midnight activities

+++ database/models.py
@@ -41,7 +41,7 @@ class Subactivity(db.Model):
     name = Column(String(100), nullable=False)
     activity_type = Column(Integer, ForeignKey('activity_types.id'))
     user_id = Column(Integer, ForeignKey("users.user_id"))
-    chat_id = Column(Integer, ForeignKey("chats.chat_id"))
+    chat_id = Column(BigInteger, ForeignKey("chats.chat_id"))
     is_removed = Column(Boolean, server_default="False", nullable=False)
